requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<0.28
python-decouple>=3.8,<3.9
psycopg2-binary>=2.9,<2.10
